+++ examples/ethersv6-ts-esm/src/index.ts
@@ -1,39 +1,77 @@
-import { wrap } from "@oasisprotocol/sapphire-paratime";
+import { wrapEthersSigner, wrapEthersProvider, wrap } from "@oasisprotocol/sapphire-paratime";
 
-import { ContractFactory, JsonRpcProvider, ethers } from "ethers";
+import { ContractFactory, JsonRpcProvider, BaseContract, Wallet, ZeroAddress } from "ethers";
 
-import { TestErc20Token } from "./TestErc20Token.js";
 import { assert } from "console";
 
+import OmnibusJSON from "../../../contracts/artifacts/contracts/tests/Omnibus.sol/Omnibus.json" assert { type: "json" };
+
+async function testTheContract(contract:BaseContract, signerAddr?:string)
+{
+    const addr = await contract.getFunction("testSignedQueries()")();
+    if( signerAddr ) {
+        assert( addr == signerAddr );
+    }
+    else {
+        assert( addr == ZeroAddress );
+    }
+
+    try {
+        await contract.getFunction("testCustomRevert()").staticCall();
+        assert(false);
+    }
+    catch(e:any) {
+        assert(e.code == 'CALL_EXCEPTION');
+        assert(e.revert.name == 'CustomError');
+        assert(e.revert.args[0] == '0x1023456789abcdef1023456789abcdef1023456789abcdef1023456789abcdef');
+    }
+}
+
 async function main () {
-    const wallet = new ethers.Wallet('0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80');
+    const wallet1 = new Wallet('0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80');
 
-    const provider = wrap(new JsonRpcProvider('http://127.0.0.1:8545'));
+    const rawProvider1 = new JsonRpcProvider('http://127.0.0.1:3000');
+    const provider1 = wrapEthersProvider(rawProvider1);
 
+    const rawProvider2 = new JsonRpcProvider('http://127.0.0.1:8545');
 
+    const signer1 = wrapEthersSigner(wallet1.connect(provider1));
 
-    const signer = wrap(wallet.connect(provider));
+    const fac = new ContractFactory(OmnibusJSON.abi, OmnibusJSON.bytecode, signer1);
 
-    const fac = new ContractFactory(TestErc20Token.abi, TestErc20Token.bytecode, signer);
-
-    console.log('Deploying');
     const contract = await fac.deploy();
-    console.log('Waiting for deployment');
     const contractReceipt = await contract.waitForDeployment();
     console.log('Deployed', contractReceipt.deploymentTransaction()?.hash);
 
-    console.log('Calling getAddress');
-    const myAddr = await signer.getAddress();
+    const signerAddr1 = await signer1.getAddress();
 
-    console.log('Calling mint');
-    const mintTx = await contract.getFunction("mint(address,uint256)")(myAddr, 100);
-    assert(mintTx.data.length > 100);  // TODO: ensure mintTx has encrypted calldata
-    console.log('mintTx', mintTx.hash);
-    await mintTx.wait();
+    // signed queries will work
+    await testTheContract(contract, signerAddr1);
 
-    console.log('Calling totalSupply');
-    const totalSupply = await contract.getFunction("totalSupply()")() as bigint;
-    assert(totalSupply === 100n);
+    const mkf = (x:any) => new ContractFactory(OmnibusJSON.abi, OmnibusJSON.bytecode, x);
+
+    // Connect wallet to rawProvider, signed queries will work
+    await testTheContract(
+        mkf(wrapEthersSigner(wallet1.connect(rawProvider1)))
+            .attach(await contract.getAddress()),
+        signerAddr1);
+
+    // Use `wrap` instead of `wrapEthersSigner`
+    await testTheContract(
+        mkf(wrap(wallet1.connect(rawProvider1)))
+            .attach(await contract.getAddress()),
+        signerAddr1);
+
+    // wrapped provider, msg.sender == ZeroAddress
+    await testTheContract(
+        mkf(provider1)
+            .attach(await contract.getAddress()));
+
+    // Switch the signer to a different provider
+    await testTheContract(
+        mkf(signer1.connect(rawProvider2))
+            .attach(await contract.getAddress()),
+        signerAddr1);
 }
 
 await main ();